+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Add BBB meeting scheduling
 - Save current language of a livesession
+- Send emails in current livesession's language
 
 ### Fixed